fix: move mqtt/m2m connect management to electron main process
After upgrade to quasar 2, the electron did not support to use MQTT/TCP
connection in render process anymore. The client will always switch to
WebSocket connection.

After some study on electron document. I decide to move the connection
management to electron main process. And export APIs from preload.js
for render process to use. This change leads to this project cannot be
run without electron, but should be not a problem.

Signed-off-by: Tim Chao <tim.chao@ingics.com>